backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dateutil==2.8.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
flask-sqlalchemy==3.1.1

